ensure all import paths are correct; refit tests; recreate docs.
