--- evaltools/plotting/redblue.py
@@ -1,3 +1,0 @@
-
-def redblue(geometries, column, specification=None):
-    pass
