+++ tests/test_data.py
@@ -37,11 +37,9 @@ def test_cvap_bgs():
     
     # Set some testing variables.
     columns = {
-        "TOTPOP19", "WHITE19", "BLACK19", "AMIN19", "ASIAN19", "NHPI19", "OTH19",
-        "2MORE19", "NHISP19", "WVAP19", "BVAP19", "AMINVAP19", "ASIANVAP19",
-        "NHPIVAP19", "OTHVAP19", "2MOREVAP19", "HVAP19", "TRACT10", "VAP19",
-        "WCVAP19", "BCVAP19", "AMINCVAP19", "ASIANCVAP19", "NHPICVAP19", "OTHCVAP19",
-        "2MORECVAP19", "NHWCVAP19", "HCVAP19", "CVAP19", "POCVAP19", "NHWVAP19"
+        "BLOCKGROUP10", "CVAP19", "NHCVAP19", "NHAICVAP19", "NHACVAP19", "NHBCVAP19",
+        "NHNHPICVAP19", "NHWCVAP19", "NHAIWCVAP19", "NHAWCVAP19", "NHBWCVAP19",
+        "NHAIBCVAP19", "NHOTHCVAP19", "HCVAP19", "POCCVAP19"
     }
     bgs = 3438
 
@@ -112,22 +110,6 @@ def test_census_tracts():
 
     assert len(data) == tracts
     assert set(list(data)) == columns
-
-
-    with gzip.open(root/"test-assignments/assignments.jsonl.gz") as g:
-        with jsonlines.Reader(g) as r:
-            counter = 0
-            assignments = []
-
-            for assignment in r:
-                if counter < size:
-                    assignments.append(assignment)
-                    counter += 1
-                else:
-                    break
-        with jsonlines.open(root/"test-assignments/test-multiple-assignments.jsonl", "w") as w:
-            for assignment in assignments:
-                w.write(assignment)
 
 def test_assignmentcompressor_compress():
     # Get the GEOIDs from the blocks.
@@ -299,5 +281,5 @@ def test_remap():
     plans = remap(plans, unitmaps)
 
 if __name__ == "__main__":
-    test_census_tracts()
+    test_cvap_bgs()
